requirements.txt
fastapi[standard]
uvicorn

sqlalchemy
asyncpg
greenlet
alembic

passlib
python-jose
argon2-cffi
python-dotenv
email-validator



python-multipart
pillow

stripe
python-dateutil

pytest
pytest-asyncio
pytest-cov
httpx

ultralytics
langchain-groq
langchain
langchain-core
langchain-community
cloudinary
reportlab
